backend/requirements.txt
# Python 3.14 Compatible Requirements
# NOTE: Python 3.14 is very new - many packages will build from source
# This will be slower and may require build tools (compilers, etc.)
# Consider using Python 3.13 for better compatibility if issues persist

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg[binary]>=3.2.10
alembic==1.14.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.0

# HTTP & Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1

# Configuration
python-dotenv>=1.0.0
pydantic>=2.12.0
pydantic-settings>=2.6.0

# Geospatial
geoalchemy2>=0.14.0
shapely>=2.0.2
geopandas>=0.14.1

# Utilities
pyyaml>=6.0.0
